chore: rename artwork import mutation files to have Mutation filename suffix
